[PROJ-29] refactor: property form 폴더구조 변경
