+++ src/pages/property/form/PropertyForm.tsx
@@ -2,9 +2,9 @@
 // import ApartListForm from '../../../components/organisms/form/ApartListForm';
 // import CarForm from '../../../components/organisms/form/CarForm';
 // import ConfirmLoan from '../../../components/organisms/form/ConfirmLoan';
-import JobForm from '../../../components/organisms/form/JobForm';
 // import SalaryForm from '../../../components/organisms/form/SalaryForm';
 import FormLayout from '../../../components/template/FormLayout';
+import JobForm from './JobForm';
 
 export default function PropertyForm() {
   return (
