+++ src/pages/property/form/SalaryForm.tsx
